requirements.txt
flask==2.0.1
opencv-python==4.5.3
pyttsx3==2.90
requests==2.26.0
RPi.GPIO==0.7.0
face_recognition==1.3.0
dlib==19.24.2
numpy==1.21.0